tests update after moving classes
